refactor: extract spending comparison chart component
